+++ tests/unit/http/response/CookiesTest.php
@@ -53,6 +53,7 @@ class CookiesTest extends TestCase
 			$this->assertArrayHasKey('domain', $cookie['options']);
 			$this->assertArrayHasKey('secure', $cookie['options']);
 			$this->assertArrayHasKey('httponly', $cookie['options']);
+			$this->assertArrayHasKey('samesite', $cookie['options']);
 		}
 	}
 
@@ -159,8 +160,9 @@ class CookiesTest extends TestCase
 		$this->assertSame('', $foo['options']['domain']);
 		$this->assertSame(false, $foo['options']['secure']);
 		$this->assertSame(false, $foo['options']['httponly']);
+		$this->assertSame('Lax', $foo['options']['samesite']);
 
-		$cookies->add('foo', 'bar', 0, ['path' => '/foo', 'domain' => 'example.org', 'secure' => true, 'httponly' => true]);
+		$cookies->add('foo', 'bar', 0, ['path' => '/foo', 'domain' => 'example.org', 'secure' => true, 'httponly' => true, 'samesite' => 'Strict']);
 
 		$foo = $cookies->all()['foo'];
 
@@ -168,6 +170,7 @@ class CookiesTest extends TestCase
 		$this->assertSame('example.org', $foo['options']['domain']);
 		$this->assertSame(true, $foo['options']['secure']);
 		$this->assertSame(true, $foo['options']['httponly']);
+		$this->assertSame('Strict', $foo['options']['samesite']);
 	}
 
 	/**
@@ -177,7 +180,7 @@ class CookiesTest extends TestCase
 	{
 		$cookies = new Cookies;
 
-		$cookies->setOptions(['path' => '/foo', 'domain' => 'example.org', 'secure' => true, 'httponly' => true]);
+		$cookies->setOptions(['path' => '/foo', 'domain' => 'example.org', 'secure' => true, 'httponly' => true, 'samesite' => 'Strict']);
 
 		$cookies->add('foo', 'bar');
 
@@ -187,6 +190,7 @@ class CookiesTest extends TestCase
 		$this->assertSame('example.org', $foo['options']['domain']);
 		$this->assertSame(true, $foo['options']['secure']);
 		$this->assertSame(true, $foo['options']['httponly']);
+		$this->assertSame('Strict', $foo['options']['samesite']);
 	}
 
 	/**