+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The major version bump is due to dropped support for PHP `7.2` and a several bre
 
 * Added `Collection::first()` method.
 * Added `Collection::last()` method.
+* Added support for the `samesite` cookie option (defaults to `Lax`).
 
 #### Changes